build: Optional CMake forwarding for build and test
This patch adds the `--cmake` option to `configure.py` and `test.py`.
This option translates options from those scripts into appropriate
options for CMake and CTest (respectively) and invokes those tools for
configuring the build and running tests.

These options should make it even easier to evaluate and test the CMake
build before we switch over entirely.

One important difference to using the scripts as they were before and
the `--cmake` option is that the `--cmake` option does not produce a
`build.ninja` file in the source directory. This means that instead of
invoking `ninja` directly in the same directory as `configure.py`, it is
necessary to invoke `ninja` like the following:

ninja -C build/release

or

cd build/release && ninja

Both scripts now log their CMake invocation for debugging,
informational, and validation purposes.

Some more details:

`configure.py
-------------

Since CMake uses an out-of-source directory for building, the script
creates the directory `build/release` for `release` mode and
`build/debug` for `debug` mode. If `--mode` is not provided, then
both directories are created and both builds are configured. It is
important to remove any existing `build` directory before running
`configure.py` with `--cmake`.

If the (CMake-only) options `--without-apps` or `--without-tests` are
included, then applications (or tests) will not be built by default
when the build tool is run.

If applications and tests are not built by default, you can still build
all applications by building the new "apps" target and build all tests
with the new "tests" target.

`test.py`
---------

Running `test.py` with the `--cmake` option will fail if tests have not
been enabled in the configure step (see above).

By default, if `--mode` is not specified, then tests for all modes are
run.

One interesting feature of CTest not yet exposed in `test.py` is running
tests concurrently with the `-j` option. You can try this by invoking
`ctest` manually instead of through `test.py`.