build: Use the same warning name for Clang and GCC
The syntax in Clang for preventing preprocessor warnings (like `#warning
Something bad might have happened...`) from producing a compilation
error looks like `-Wno-error=#warnings`. The syntax in GCC for the same
thing looks like `-Wno-error=cpp`. The '#' character confuses come CMake
modules, even when it is escaped as in '\#'.

Clang added an alias to the `#warnings` syntax to match the GCC flag in
commit e397992a6204455301c7d67d25025422eeec98a0 (based on the Git mirror
of the Subversion repository), which is present in the 5.0 release
series of LLVM.

Also, the flag was mispelled. It should have been `#warnings` and not
`#warning`.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <66a373c272a42cafbcf30e77541d31401379f9ce.1525290560.git.jhaberku@scylladb.com>