be more descriptive with failed syscalls for filesystem operations
We currently return a system_error with an error code when any syscall
goes wrong. That is helpful in some situations but in many others we
would like to use other exceptions with more descriptive error
messages.

This is particularly true for the case of filesystem operations, for
which knowing which path fails can go a long way in debugging the
issues.

Tests: unit (release)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180509215921.11224-1-glauber@scylladb.com>