reactor: fix calculation of idle ticks
When we go to sleep we add ticks to the block detector artificially,
since it missed some. But we are doing it wrong and adding ticks as
(start - end), where in reality the real delta should be (end -start).

I found this by code review. It is not clear to us what the effects of
this are. But let's fix this for consistency.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180513102718.2429-1-glauber@scylladb.com>