reactor: allow users to compute steal time
There are many situations in which external processes steal CPU time
from seastar application. Those situation can cause a lot of havoc,
since specially if not in overprovisioned mode seastar applications
expect to have the CPU by themselves but it is very hard to see.

If we had metrics for that, it would be easy to see that this is what's
happening.

We can do that in an efficient way by using
clock_gettime(CLOCK_THREAD_CPUTIME_ID). That will give us a measurement
of how much time we actually spent running this thread. We can then
compare that to how much time we thought we ran for. The delta between
them is steal time-- very similar in concept to steal time in
hypervisors like KVM and Xen.

Note that we can't use the current cpu_busy_ns time since it discounts
idle time. But idle time, from the point of view of seastar is time
during which the reactor thread actually ran. We need a metric that
discounts only the time during which we actually slept.

Signed-off-by: Glauber Costa <glauber@scylladb.com>