json: json_path to support in path parameters
In path parameters are parameters that can be found anywhere in the
path.

For example:
/mypath/{param1}/morepath

The http server supports matching rule that are created from a combination of
parameters and fix URL strings.

This patch changes the json_path to support path parameters.

In json_path (the class that describe a path and is used by the code
generator) instead of a bool that mark the type of path param we now
use an enum and instead of tuple a struct path_part that describe one
part of the define path, that can be either a parameter or a static
string.

The add_alias method in route was changed to support the different
alternatives.

In the above example, the json_path will be compose of 3 parts.
The path part will be:
/mypath

and two path_part: a parameter named 'param1' and a fixed string
'morepath'.