iotune: be resilient against sys files with different formats
We expect scheduler files for sysfs to look like this in our
block devices:

 foo bar [quux] blah

In reality, some of them, where only one scheduler is present will
not use the [].

Be resilient against that by checking the result of the find operations.
If one of the characters is not found, just use the entire string. If
the one available scheduler is still none or noop we can treat it the
same way and be silent.

Fixes #449

Signed-off-by: Glauber Costa <glauber@scylladb.com>