utils: make make_visitor() public
Currently, seastar::visit() is limited to variants compatible with
boost::apply_visitor() which it uses internally.

However, a helper function make_visitor() is much flexible as it just
creates a visitor object from the provided function objects. This makes
it a viable helper in any situation when some kind of variant needs to
be visited.

Message-Id: <20180520134814.6752-1-pdziepak@scylladb.com>