io_queue: store the I/O Queue pointer in the file object
For file objects that go through I/O Queue, we can store the pointer in
the implementation itself. In this patch, we just copy the pointer at
object creation time, so we don't gain much. But what this will allow us
to do is to choose from one of many options when we support multiple I/O
queues per system.

Signed-off-by: Glauber Costa <glauber@scylladb.com>