io_queue: make queue_request into instance method.
queue_request is currently a static method of the I/O Queue. The reason
for that is that we may do remote dispatch to a foreign I/O Queue, so it
has been deemed historically safer to just follow the local pointer of
the remote queue once we are already in the right shard.

However, every reactor has a naked pointer pointing to the I/O queue
where its requests should be dispatched. Therefore it is safe to make
it into an instance member.

This doesn't matter much now, but will matter when we start having
multiple I/O queues in the system - as we will be capable of routing
the request to the right one.

Signed-off-by: Glauber Costa <glauber@scylladb.com>