Log stack trace during exception throwing
When C++ exception is caught the information about where it was
thrown is already lost, so sometimes it is hard to know what code exactly
threw it. This patch adds an ability to print a stack trace during
exception throwing which may help debugging. The logging is done by new
"exception logger at trace level, so it can be enabled and disabled
dynamically.