IOTune: allow duration to be configurable and change its defaults
Results for read benchmarks can vary wildly if the data we're reading
from is too small and can fit in the device's cache. We know that, which
is the reason we write for way longer than we read.

Ideally, we would write for a very long time. But this is at odds with
our other goal of providing a good UX, where the user don't have to wait
for a long time for the test to complete.

However recently we found out that our defaults (60s total) is not long
enough for the largest boxes in the AWS I3 class. Results are distorted,
and may appear faster than they actually are.

This patch moves the total write time to 120, which I verified
experimentally to provide results that align with the truth for AWS
i3.metal-- here "the truth" meaning the results I got by doing 10-minute
writes.

The storage for i3.metal is so fast that if it works well there, it
should work well everywhere. But still, I am bringing back the duration
option that iotune1. The main difference is that while for iotune1 the
duration represented the time over which we would give up and timeout,
for iotune2 it represents the total time of the test.

We will keep the same proportions: 70% of the time will be spent in the
sequential writes, because that is the test that will lay down the files
that the other tests will use. The remaining 3 other tests will use 10%
of the time each.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180622203640.9796-1-glauber@scylladb.com>