temporary_buffer: Add clone method()
Copying is prohibited for temporary_buffer but sometimes it comes in
handy to create a deep copy of a buffer.
Before this fis, this was only possible with a complex syntax like:
    temporary_buffer<T> buf;
    temporary_buffer<T> buf2{buf.get(), buf.size()}

But if you need to do that several times, this quickly becomes noisy and
asks to be factored out.

The clone() method covers this case and fits in naturally into
temporary_buffer design.

Signed-off-by: Vladimir Krivopalov <vladimir@scylladb.com>
Message-Id: <6e110a524cce2e620f7bff98f7d23d83c86ed4e8.1530667807.git.vladimir@scylladb.com>