reactor: pollable_fd: limit fragment count to IOV_MAX
Requesting to send more than IOV_MAX fragments results in an EMSGSIZE
error. write_some() is not required to send all data, so it is
sufficient to limit the number of segments to no greater than IOV_MAX.

A minor performance issue is that when the number of fragments gets
limited write_some() will never speculate that the file descriptior can
still accept more data. However, the average size of a fragment is
expected to be at least several kilobytes, which means that sendmsg()
was asked to send several megabytes (assuming IOV_MAX = 1024), so the
lack of speculation is unlikely to become a problem.

Refs #361.

Message-Id: <20180709132129.21786-1-pdziepak@scylladb.com>