tutorial: update build as needed for newer pandoc
Unfortunately, pandoc 2 (which is available, for example, on Fedora 28)
broke some old features and changed formatting details a bit, so I needed
to modify the pandoc options, and also to change the HTML splitting code and
the stylesheet to support both old and new versions (I tested Pandoc 1.19
and Pandoc 2.0.6).

I moved the commands to convert MD to HTML and PDF, which have now grown
even longer, to separate scripts - doc/md2html and doc/md2pdf.
These scripts can also be used for the Cmake-based build, so we
don't need to repeat these commands.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180730145545.12397-1-nyh@scylladb.com>