seastar-addr2line: optimize the case where stdin is not a TTY
When stdin is not a TTY (but e.g. a pipe) there is no need to read the
input char by char. There will be a proper EOF when the input is
exhausted and therefour we can use sys.stdin directly in this case, no
need for the wrapper that improves the user experience when pasting
traces manually on the terminal.