seastar-addr2line: allow user to control the amount of context printed
Non-backtrace lines are mostly just noise and have no value. Provide an
option to the user to not print them at all or print just a limited
number (controlled by the user) of those that preceed the backtrace, for
identification purposes. This is controlled via the new --before|-b
command line argument.
To further help identifying backtraces assign an id to each backtrace
and print it before printing the backtrace. The id is simply a counter
that starts from 0 and is incremented at each seen backtrace. This will
be even more important in the future as we plan to not print already
seen backtraces and instead just print a backreference to the place
where they were already resolved.
Also make sure backtraces are separated backtraces with an empty line.