posix-stack: Choose local port based on the shard number connect() runs on
If local address provided to connect() has wildcard port try to use
a port number that satisfy equation: port modulo number_of_shards =
current_shard.  If a server uses port based load balancing and has
the same amount of shards as a client it will result in connection
going to the same shard number in the server, otherwise no harm will be
done. Connecting to the same shard number is beneficial for applications
that distribute work based on a hash of the data been processed since
in such systems this will eliminate internal cross-cpu hop from the
pipeline.