seastar-addr2line: fix redirecting output to a file
Pipe addr2line's stout to the script and print it from there to avoid
problems arising from buffering when the script's output is redirected
to a file.
This fixes the issue where, when stoud is redirected to a file, the file
would contain first all the output from the addr2line calls and then
those from the script itself.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <f6ff430c72779342b1176bbf98d2a3c57e4c9507.1533895164.git.bdenes@scylladb.com>