Do not use `constexpr` closure in function body
According to cppreference.com [1], a `constexpr` variable must, among
other things, be of a `LiteralType`.

In C++14, a closure type (produced by a lambda expression) is not a
`LiteralType`, but as of C++17 it is.

Therefore, for compatibility, we can declare this variable as `const`
without any loss of functionality: we don't need the closure to be a
compile-time value.

[1] https://en.cppreference.com/w/cpp/language/constexpr