Prevent copy during initialization in C++14
In C++14 [1], copy elision in expressions such as

   T x = T(arg1, arg2);

was optional, such that rather than invoking `T`'s ctor and then its
copy-ctor or move-ctor, the compiler only need to invoke the ctor.

This changed in C++17, so that copy elision is mandatory.

This means in practice that expressions like

     auto x = my_type(arg1, arg2);

are effectively equivalent to

    my_type x(arg1, arg2);

For compatibility with C++14, we need to use the second form in this
file because the `client` class is not moveable nor copyable.

[1] https://en.cppreference.com/w/cpp/language/copy_elision