tests: Ensure test thread is always joined
In some instances, I observed that the destructor of the static
`test_runner` instance did not run when a program linking against the
testing library terminated.

The rules around when destructors of static objects can invoked are
somewhat complex, so to avoid ambiguity and ensure that the thread is
always joined, we move the bulk of the destructor to a `finalize`
function and invoke it manually.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <a57f361c8131d14f300cd234bb25e57a8eafe3d7.1534532050.git.jhaberku@scylladb.com>