reactor: define max_aio_per_queue for C++14
With the switch to CMake, compiling Seastar according to the C++14
standard uncovered linker errors of this form:

    libseastar.a(reactor.cc.o): In function `seastar::io_queue::make_fair_queue_config(seastar::io_queue::config)':
    /home/jhaberku/src/seastar/build/../src/core/reactor.cc:1153: undefined reference to `seastar::reactor::max_aio_per_queue'
    libseastar.a(reactor.cc.o): In function `seastar::disk_config_params::parse_config(boost::program_options::variables_map&)':
    /home/jhaberku/src/seastar/build/../src/core/reactor.cc:4052: undefined reference to `seastar::reactor::max_queues'

The cause of these errors is a change in how C++17 treats constant
static members of classes compared to C++14.

In C++14 [1], a `static constexpr` member of a `LiteralType` must be
initialized with a braced and constant expression like

    struct foo {
        static constexpr int my_constant{42};
    };

However, if the member is "odr-used", then a declaration at namespace
scope is still required, despite the definition in the class:

In `foo.cc`:

   constexpr int foo::my_constant;

The phrase "odr-used" has a precises meaning, but a rough rule-of-thumb
is that a member is "odr-used" if it's ever treated like a variable and
not just a constant expression (for example, the member's address is
taken).

In C++17, new "inline" static members mean that a separate declaration
is not required, and any out-of-line declarations/definitions are
deprecated but still supported.

Thus, to fix the linker errors, we provide declarations for
`reactor::max_queues` and `reactor::max_aio_per_queue`.

[1] https://en.cppreference.com/w/cpp/language/static

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <f16ec2d255813b9c8e23002773765e476cc0bc13.1534532050.git.jhaberku@scylladb.com>