seastar-addr2line: add --verbose option
In verbose mode, in addition to the resolved symbols, the originating
module and the resolved address is printed. This is helpful as function
names don't always give away which module the frame originates from.
It's even more useful when some addresses cannot be resolved and
just a very unhelpful "??" is printed instead of the resolved symbol.
Also useful if the user has other means to resolve addresses that the
script failed to resolve.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <78f3021ffd6a782c743a38e02d697ddf77b5c773.1535026792.git.bdenes@scylladb.com>