future: make available() const
This is, both conceptually and practically, a const method.
This change also allows to filter a range of futures based on whether
they are available. This requires a const method as some filter
algorithms (e.g.  boost::adaptors::filtered) pass a const& to the
filtered object into the filter function.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <38bc96af39258a35bf5643fdbf0f4195cdcfae88.1535435789.git.bdenes@scylladb.com>