build: use 'ppa:scylladb/ppa' instead of URL for sourceline
in hope to help travis CI to import the APT key of the PPA repo

the travis CI stared to fail recently due to unauthenticated ppa
repo used for installing seastar depedencies. to address this, we
need to use the 'ppa:scylladb/ppa' instead its URL for sourceline.
so travis CI can use apt-add-repository instead of adding the sourceline
to /etc/apt/sources.list directly. the former method will import the
key for us, and hence address the "unauthenticated packages" issue.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20180904041341.37481-2-tchaikov@gmail.com>