seastar-addr2line: don't invoke addr2line multiple times
The current script invokes the addr2line binary every time it sees a new
backtrace. This is terribly inefficient. The biggest problem here is not
just the fact that we have to fork a new process, but the fact that
every new invocation has to reload the entire symbol table.

We'll fix this by opening one addr2line process per module we're dealing
with, and then using pipes to communicate with it. As we send an input
line, it will respond with the corresponding output line.

To test this, I fed a file with 7 backtraces to the script before and
after. Aside from verifying that the output is still the same, I timed
the difference to:

Before:
real	1m50.030s
user	1m38.046s
sys	0m11.795s

After:
real	0m6.871s
user	0m0.032s
sys	0m0.020s

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180913222618.18162-1-glauber@scylladb.com>