tests/perf: introduce abstract result_printer
So far, test results were only printed to the stdout. This is about to change
and an option to write them to a json file is going to be added. An abstract
class result_printer is added that all implementation of result handling
(i.e. both stdout and json) can derive from.