rpc: introduce lz4_fragmented_compressor
A flaw of lz4_compressor is that it needs to allocate contiguous
input and output buffers for the whole message. Apart from the
cost of linearisation this will also put additional pressure on
the memory allocator which may struggle to provide large
contiguous buffers.

One solution may be to reuse those large buffers, but that means
both compressed and decompressed data would have to be copied
to/from those large buffers. Moreover, a single call to lz4 functions
cannot be pre-empted and may take unacceptably long time.

The LZ4_FRAGMENTED compressor addresses those problems and while
more complex than the naive approach allows compression and
decompression without the need to allocate any large contiguous
buffers. Internally, it uses 32 kB chunks and tries to compress and
decompress them without any additional memory copies.

Under the hood LZ4_FRAGMENTED uses lz4 stream interface with custom
metadata (4 bytes per 32 kB of uncompressed data). This allows
compressing data in chunks without sacrificing compression ratio
since compression of the current chunk depends on the previous one
(i.e. they are not independent from each other and need to be
decompressed in order, which is not a problem in this case).
Unfortunately, the stream interface appears to be slower than the
default one and the loss of performance depends largely on the size of
the chunk.