fstream: remove default extent allocation hint
We set the default extent allocation hint to 1MB, in order to reduce
fragmentation. However, if the disk happens to be a RAID array, XFS attempts
to align that extent [1]. So, those small files, while they only occupy the
space they use, they prevent use of the remaining space after them until the
next stripe alignment. As a result, you can see ENOSPC errors even if the disk
is mostly free - all the aligned blocks are allocated, and XFS won't allocate
the unaligned blocks.

To prevent this, remove the default extent allocation hint.  Instead, increase
the default file stream output buffer size to 64k, so that extents will be at
least that large.

Ref scylladb/scylla#3867.

[1] https://www.spinics.net/lists/linux-xfs/msg22933.html
Message-Id: <20181021092704.15790-1-avi@scylladb.com>