io_queue: only increment statistics when request is admitted
The graphs we generate out of our bandwidth and IOPS stats can be a bit
misleading. This is because we increment those quantities as soon as the
request is generated, and not when it is admitted.

It is, therefore, measuring bandwidth and IOPS generated by the consumer
and not bandwidth and IOPS that the I/O Scheduler is generating. In contended
scenarios those things can be quite different

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20181022201510.26962-1-glauber@scylladb.com>