tests: Kill only us in `signal_test`
When the process ID (PID) argument to `kill` is 0, the effect is to send
the signal to all processes of the same process-group as this process.

When a unit test is executed by another process (the "runner"), it
shares the same process group. Therefore, the signal terminates the
runner as well.

This change ensures that only the current process receives the signal.
We also include the necessary headers for `getpid`.

Fixes #518