build: Remove experimental CMake support
The previous support for CMake was experimental, and these experiments
enabled significant learning and development of CMake expertise.

Now that we are older and wiser, it is clear that the previous approach
was fundamentally ineffective in many respects. It is best to start from
scratch, incorporating all the lessons we've learned and discarding all
the baggage we've collected.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <d4875885c222002187365d967f89b9e29ff89b1e.1540591209.git.jhaberku@scylladb.com>