print: add fmt_print()
Because fmt's printf() emulation changed subtly (format("%s", 5) is now
broken) we want to wean ourselves off printf style and move to the new
form, which is set to ve standardized. As an intermediate meaure introduce
fmt_print(ostream, fmt, ...), same as fmt::print(ostream, fmt, ...) except
that it also returns ostream, which is expected by some of the existing code.
This asists mechanical translation.
Message-Id: <20181027202753.18515-1-avi@scylladb.com>