prometheus: pass names as query parameter instead of part of the URL
We have infrastructure to parse a name parameter in the prometheus URL,
and upon finding one, return only metrics that have that start with that
string.

However, this today relies on url name parsing instead of query parameters.
Prometheus has a way in its configuration file to generate sliced queries
based on query parameters so it would benefit us to switch to that.

Fixes #526

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20181031180626.5992-1-glauber@scylladb.com>