core: Fix build for `gnu++14`
The declaration and definition of `s_used_scheduling_group_ids_bitmap`
previously had the form of "constant initialization" [1] without a
constructor call, which means that it was "value initialized" [2]. This
means that the implicit constructor for `std::atomic` was invoked,
producing a temporary value, which was supplied to the copy ctor of
`std::atomic`.

In C++14, copy elision is not required, and `std::atomic` is not
copyable. This resulted in a compilation error. In C++17, copy elision is
required [3], and the move ctor of `std::atomic` is invoked.

To fix the issue, we use "direct initialization" [4], which ensures that
we invoke the ctor of `std::atomic` directly instead of producing a
temporary.

Fixes #527

[1] https://en.cppreference.com/w/cpp/language/constant_initialization

[2] https://en.cppreference.com/w/cpp/language/value_initialization

[3] https://en.cppreference.com/w/cpp/language/copy_elision

[4] https://en.cppreference.com/w/cpp/language/direct_initialization

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <df4c31630e61190322ed849f8bdf3053308ab4eb.1541095276.git.jhaberku@scylladb.com>