prometheus: replace the instance label with an optional one
Modifying the instance label is condemn in the prometheus eco system and
is generally accepted as should be done by the prometheus server.

If for some reason users do wants to add their own custom label (which
they probably shouldn't), this patch adds the ability to set a custom
label with the addition of optional label_instance.

By default this label will be emitted.

Fixes #477

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20181115083118.24233-1-amnon@scylladb.com>