build: drop bundled c-ares package
Seastar depends on a new c-ares feature, virtualized socket functions,
so it carried a c-ares submodule to make it easier to consume. That feature
has now been available in c-ares 1.13 for over a year, and is available
in current distributions, so we can drop the submodule and depend on the
distributed c-ares package.

Users of old distributions will have to build c-ares themselves and
provide it on the include path.

Tests: unit (release) + dns_test (which has one preexiting failure in tcp_test)

Note that dns_test is not part of test.py.
Message-Id: <20181115132831.24047-1-avi@scylladb.com>