build: Fix protobuf generation rules
protoc 3.3 seems to handle paths differently than later releases. This
change modifies the way protoc is invoked to be compatible with all
likely versions.

We want the generated `.pb.h` and `.pb.cc` files to placed in a
directory like

    build/release/gen/src/protobuf

and we want the `.pb.cc` file to have imports of the form

    #include "metrics2.pb.h"

protoc's `-I` argument instructs protoc how to generate the import
statements.

Therefore, we invoke protoc with a qualified import path and an input
file which includes that path:

    protoc --cpp_out=build/release/gen/src/proto -Isrc/proto src/proto/metrics2.proto

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <6625b60c296d3f34e8c7d0e60847aaf68675989d.1542750394.git.jhaberku@scylladb.com>