reactor: change cpu stall detector not to depend on timer thread
The reactor will soon gain an aio poll mode, which does not have a timer thread.
To avoid losing the stall detector, we migrate it to use a timer+signal directly.

In order to prevent excessive signals, the timer is set to the report threshold,
rather than the task quota. This means that almost every delivered signal
corresponds to a quota violation (the exceptions are when we've stopped running tasks
by the time the signal was delivered).

To avoid excessive calls to timer_settime(), we allow ourselves some slack in the
timing. For a threshold of 10ms we'll set the timer to 13ms and rearm it after 3ms
have gone by. So when we enter task execution, we have between 10ms and 13ms on the
clock.