seastar-addr2line: be robust against missing libraries
If a library doesn't exist in a particular path, addr2line will just
return an IOError, so we need to be robust against that.

It would be cleaner to already know that in advance in the constructor,
but unfortunately the process won't really exit with a return code until
we try to pipe something to it - so we wrap the decoding code in a
try/except block.

Now, instead of an exception we print the address and library name and
move on.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20181129200053.8116-1-glauber@scylladb.com>