io_scheduler: allow nodes to boot with fewer I/O queues than NUMA nodes
We have been historically mandating that each NUMA node should have its
own I/O Queue, for performance reasons (so we don't have to copy
pointers across NUMA nodes).

However, if the devices are slow enough, we don't really have much of a
choice. To accomodate for that, we will prefer to allocate everything
locally to the NUMA node, but if it is not possible we allocate shards
beloning to NUMA nodes without their own coordinators to random nodes
among the ones that exist.

I believe that assigning in a round-robin fashion is better at this
point to try to play proximity tricks: first, in the common case there
will be just a single NUMA node with a coordinator anyway. Second, if
there are more and we try to find the closest we may end up with
coordinators that have way more shards than some others.

Fixes #500

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20181203145721.20414-1-glauber@scylladb.com>