build: Correctly forward cflags to CMake
The value of `Seastar_CXX_FLAGS` must be a semicolon-separated list. Not
only was the value of `user_cflags` not forwarded in this format to
CMake, but it was populating a non-existent CMake variable.

Additionally, we now correctly forward the value of `user_optflags` to
CMake when we're building the "release" mode.