Use a dummy line as a separator in seastar-addr2line
addr2line can produce 1 or more lines for each line of
input. Currently seastar-addr2line doesn't account for that and can
miss some output.

It cannot simply try reading more, since it never knows when addr2line
is done.

One option would be to run addr2line once per address, but that is way
too slow.

Another is to feed addr2line all the addresses we want first and do a
second pass over the data, which is undesirable.

It is not sufficient to do a second pass over each backtrace, since it
cannot detect the end of the last address.

Yet another option is to have send addresses from one thread and print
stacks from another. This works and is a bit faster, but is somewhat
complicated in the current code structure.

What this patch does is send an extra dummy line so that there is a
marker to look for in the output.

The patch also uses the opportunity to simplify the code a bit:

* We know when '\n' are present.
* No need to batch addresses since addr2line is slow just to start.

A file with 10K addresses goes from 4.015s to 4.150s.

We can use threads if we need to speed this up, but even then the
dummy line is a good thing since it guarantees the printing thread
never blocks for too long on the last address of a given module.

Thanks to Tomasz for the idea.

Reviewed-by: Botond Denes <bdenes@scylladb.com>
Message-Id: <20181206150842.4945-1-espindola@scylladb.com>