weak_ptr: fix pointer semantics
C++ makes a distinction between const pointers and pointers to const.
The former does not allow the value of the pointer to be changed, but
allows non-const operation on the pointed-to object. To reflect this,
the `*` and `->` operators are const methods in C++ smart pointers.
Follow suit and drop the (incorrect) non-const overloads of said
operators in weak_ptr.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <0fb6c4deeb2740be8b0349da8c315524323512b0.1544181442.git.bdenes@scylladb.com>