tests: Restructure library files
This change reorganizes the files comprising the unit testing library in
anticipation of them being a part of the public interface of Seastar.

The definition of `SEASTAR_TEST_CASE` lives in `test_case.hh`.

The definition of `SEASTAR_THREAD_TEST_CASE` lives in
`thread_test_case.hh`.

`test-utils.{hh,cc}` is now `seastar_test.{hh,cc}` to correspond to the
name of the class.