Export the testing library publicly
This change allows consumers of Seastar (libraries or programs) to
define and execute tests running in asynchronous Seastar contexts. The
test library is built on Boost Test and previously existed only for
Seastar's internal use.

We define the CMake target `Seastar::seastar_testing` and a pkg-config
library called `seastar-testing`.

Consumption of these libraries (and a basic example) is tested in
`tests/dist` and executed as part of the `test_dist` target.