net: udp: add shutdown_input()/shutdown_output()
Currently, udp has a synchronous close() method and no
shutdown_input()/shutdown_output() like tcp. This happens to work because,
for the posix stack, close() is able to remove all kernel state immediately.
The native stack handles this in another way, by lazily destroying the state
after close().

Change that have shutdown() methods that are used to trigger failures in
pending reads and writes, and require the caller to wait on those pending
reads and writes before calling close(). This is important because when
we move to IOCB_CMD_POLL, we won't have a synchronous way of removing
kernel state when we want to close a connection - both io_cancel() and
shutdown() will respond in the next io_getevents(), not immediately.

While we could have taken ownership of the state and removed it lazily,
that is bad for resource control, and is against the spirit of Seastar
that requires the caller to manage object lifetime. So this patch
introduces shutdown_input() and shutdown_output() in udp_channel, and
requires users of recv() and send() to call them in order to terminate
pending operations.
Message-Id: <20181216185847.31882-1-avi@scylladb.com>