Use the same compiler flags for demos, tests and apps
There are a few flags we want to use in seastar but not force on every
user. Currently they are used only when building libseastar.a, but not
any of the executables.

This patch adds variables for holding defines, libraries and flags,
which the seastar library then uses as PRIVATE.

The patch also adds a convenience seastar_with_flags that includes
both seastar and the variables publicly. It is then used with
executables we want to use the same compiler flags as seastar.

This doesn't change the flags used or if they are public or
private. Each of those will be an independent discussion.

Fixes #544

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Reviewed-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <20190109023756.31070-1-espindola@scylladb.com>