reactor: move pollable_fd and related classes to an internal header files
pollable_fd and friends are internal data structures and should not be exposed.
This is still too difficult at the moment because public classes depend on them,
but this patch starts the work and moves them to a separate header file. It also
makes it possible to make syscall_work_queue, which depends on pollable_fd,
private.