reactor: fix loss of highres timers with the epoll backend
The lowres_timer_pollfn and signal_pollfn pollers have a hidden
dependency: lowres_timer_pollfn::try_enter_interrupt_mode() may arm a
timer, which (if it happens to be the first highres timer to be armed)
register a signal handler. But signal_pollfn assumes there are no changes
to signal registration between signal_pollfn::try_enter_interrupt_mode()
and signal_pollfn::exit_interrupt_mode(); if there are, the newly registered
signal becomes blocked.

This was broken in 5aea50a8a4f46316 ("reactor: virtualize high resolution
timer processing") and is seen in memcached tests with the epoll backend.

Fix by reversing the order of poller registration so that the problem is
never triggered. The bug deserves a better fix, but let's start with
something simple.
Message-Id: <20190112202337.22987-1-avi@scylladb.com>