tutorial.md: also give a direct compilation command, not just cmake
The Seastar tutorial in its getting started section used to have a simple
one-line compilation command using c++ and pkg-config. This was replaced
by a more elborate cmake example. I think it was a mistake to replace the
example instead of adding it. The cmake example is unusable for people
who have no intention of using cmake, and moreover does not work (unless
I'm misunderstanding something) against an uninstalled Seastar.

So in this patch I add also the original pkg-config example, with
compilation order fixed because we no longer use "--whole-archive"
so need the library to appear *after* the application code.

Fixes #565

Note that the compilation command suggested here doesn't fully work
today - issue #579 describes a problem with "--static" and there's
a problem with "fmt" being compiled as a shared library but "--rpath"
not added. But these can be fixed incrementally in later patches.
We need the document to at least say what *should* work.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190114111823.9808-1-nyh@scylladb.com>