build: use __has_include to detect std::filesystem header
and the <filesystem> header is available (starting gcc 8).

Note that testing __has_include alone is insufficient since
the header may be available while compiling with an older dialect,
but it contents is included only for c++17 and later.

Use the __cpp_lib_filesystem and __cpp_lib_experimental_filesystem
feature macros to define the filesystem namespace.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>