core: output_stream: add fast path for small writes
output_stream may be used in various ways. It provides zero-copy and
buffered interfaces. The latter may be used as a mix of large and small
writes. For writers that perform multiple small writes to the
output_stream it is beneficial to add a fast path for cases where the
whole write fits in the current buffer.

Message-Id: <20190125155954.5140-1-pdziepak@scylladb.com>