future: propagate broken_promise exception to abandoned continuations
With futures and promises model it is possible that a promise is
destroyed before ever having a value set. In particular this may lead to
two problematic scenarios:
 - promise abandons a future, and then there is an attempt to attach a
   continuation or a thread to that future
 - promise abandons a future to which a continuation or thread is
   already attached
At the moment doing so leads to an undefined behaviour varying from the
continuation chain getting stuck to a segmentation fault. This patch
changes that so that in both of these cases a broken_promise exception
is propagated to the waiting continuation or thread thus avoiding
undefined behaviour.

Message-Id: <20190129202023.14187-1-pdziepak@scylladb.com>