build: Fix the `StdFilesystem` module
When a client uses CMake to link against Seastar, CMake will process the
body of `FindStdFilesystem.cmake` without processing Seastar's
`CMakeLists.txt`. Therefore, we cannot use Seastar CMake cache variables
in the definition of `FindStdFilesystem`.

We define a new variable which can be set prior to invoking

    find_package (StdFilesystem)

in order to control the C++ dialect used in its tests. If this variable
is not set, we use the compiler's default dialect in the test.

When Seastar is configured, it sets `StdFilesystem_CXX_DIALECT` to the
value of `Seastar_CXX_DIALECT` so that these two variables are
synchronized.

Fixes #590

Tests: dist (release)
Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <26a7696e453c73d596a997710c5c537a279df6d0.1548896063.git.jhaberku@scylladb.com>