rpc: support closing streaming when only sink or source was created
Currently to close a stream both sink and source have to be closed, but
sometimes only one is available. It may happen that rpc that send a sink
to a server fails, so the corresponding source is never received and
when that happen there is no way to close exiting stream.

The patch fixes it by marking each side of stream as open at the time sink or
source is created. If one of the sides was never created it should not
be closed.

Message-Id: <20190131091358.GB3172@scylladb.com>