Add an option for building with split dwarf
For now the main advantage is that --gdb-index speeds up gdb
startup. On my machine running gdb on a debug version of scylla with
the commands

b main
r
q

with Seastar_SPLIT_DWARF set to ON takes 17.5 seconds, with it OFF it
takes 45.1 seconds.

Leaving some of the debug info in .dwo files should also make it
viable to drop -gz, which should save more time by not compressing and
decompressing debug info.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190201193144.40221-1-espindola@scylladb.com>