net: ipv6: fix ipv6 detection and tests by binding to loopback
On my workstation, binding to "::" (the default ipv6 address) results in
local_address() returning "::" as well. This causes sendto() to fail with
a network unreachable error, failing the test.

In a docker container, the bind to "::" succeeds even though the loopback address
does not have an ipv6 address, causing the test to fail.

Fix by binding to "::1", both in ipv6 detection and in ipv6 tests. I verified
(via strace) that detection works on an ipv6-enabled machine.
Message-Id: <20190206170318.7756-1-avi@scylladb.com>