build: switch from travis-ci to circleci
travis-ci does not have first-class support for docker, and its supported images
are very old (>2 years). Switch to circleci which allows building in a docker
container directly, without intermediate steps in a VM.

The docker image is names scylladb/seastar-toolchain and was generated from
docker/dev/Dockerfile.

No attempt was made to optimize the build or have a wide test matrix; the intent
is to provide a working foundation that can be optimized later.
Message-Id: <20190206181051.11128-1-avi@scylladb.com>