build: Make the use of sanitizers transitive
While some sanitizers do not have to be enabled for programs which link
against libraries with sanitizers enabled, we can be conservative and
ensure that sanitizers are enabled transitively all the time.

This was already the way that pkg-config worked, and this change ensures
that consumers of Seastar using CMake behave the same way.

We can eliminate the `sanitizers_cflags` pkg-config variable because the
sanitizers are part of the expansion of `INTERFACE_COMPILE_OPTIONS`.

We can also eliminate the `Seastar_DEVELOPMENT_LIBS` variable.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <fabd80a995ac280345a501614512c05d669bddd8.1549571398.git.jhaberku@scylladb.com>