future-util: optimize parallel_for_each() with single element
If parallel_for_each is given a single element (or more than one, but with just
the last one not immediately available), then we can avoid creating any state
and return that last future. This not only saves state creation, but also the
continuation for then_wrapped().

Tests: unit (release)
Message-Id: <20190210173514.11276-1-avi@scylladb.com>