reactor: Fix an infinite loop caused the by high resolution timer not being monitored
In request_preemption we have a loop waiting for need_preempt to
return true. For that to happen, the kernel must update the
preemption ring.

The ring is updated any time a request finishes, but
request_preemption cannot assume that there is anything left to do. It
must ensure that timerfd is in the preemption ring.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190212162749.9967-1-espindola@scylladb.com>