rpc: fix insertion of server connections into server's container
Server hold shared pointer to all its connection which ensures that it
can stop and wait for all of them when terminated and that connection
will not disappear prematurely since there is always one shared pointer
to it. But if streaming is not enabled for a server the insertion code
does not work: it inserts only one connection and drops all others
because they all have the same id. Fix it by generating unique id for
all connection even if streaming is not enabled.

Message-Id: <20190214083523.GA16392@scylladb.com>