docs: Ensure all header files are searched by Doxygen
There are header files which do not contain Doxygen mark-up (i.e., are
not documented) but which contain definitions which need to be processed
by Doxygen in order to successfully parse header files which *are*
documented.

With this change, the rendered Doxygen HTML includes previously-missing
documentation for classes like `seastar::future`.

Fixes #605