net: fix tcp load balancer accounting leak while moving socket to other shard
The connection load balancing has two paths: one where the non-listening (ap)
shard is already waiting to a connection, and another where the listening
(main) shard got a connection first, and is queueing it until such time as
the non-listening shard gets a round to dequeue it. In the second case, we
forgot to pass along the conntrack handle, and so we did not account for it
in the load balancer data structures. This can cause a shard which used the second
path to later receive more than its fair share.

Tests: unit(dev), some manual tests with httpd.
Message-Id: <20190223153115.6405-1-avi@scylladb.com>