Fix C++-17-only uses of static_assert() with a single parameter.
static_assert was introduced in C++11 as taking two arguments - the
expression to check, and an error message. Only C++17 added the option
of dropping the second argument (and just showing which expression failed),
but Seastar is supposed to be usable for C++14, so we shouldn't use this
feature.

I found three occurances of single-argument static_assert() that crept into
our code based, and fixed them in this patch.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190224131712.1059-1-nyh@scylladb.com>