Fix CLI help message (network & DPDK options)
Currently the CLI help message never shows the network stack options or
DPDK settings. This is because the help message is generated once by
Boost::program_options in the constructor of app-template, before
smp::configure is called. This patch separates the reactor configuration
routine from the registration of network stacks and their respective
CLI options.

Message-Id: <20190225081355.13749-1-hippolyte.barraud@gmail.com>