tests/perf: support multiple iterations in a single test run
At the moment, the PERF_TEST body is expected to perform a single
iteration of the tested operation. However, if the operation is very
cheap then it may be more desirable to perform multiple iterations in a
single run of the test. This would allow to reduce any overhead and also
makes it possible for the user to choose whether they want to measure
the latency or throughput.

This patch adds that option. If the PERF_TEST returns a size_t or
future<size_t> that value is interpreted as the number of iterations run
in a one call to PERF_TEST body and the final results are adjusted
accordingly.
Message-Id: <20190304125745.31943-1-pdziepak@scylladb.com>