core: show address in error message for posix_listen failures
posix_listen may fail, usually when the port is already in use.
The error message it throws is quite cryptic, though, in the sense
that it doesn't tell us which address we failed to bind to.

This patch enriches the exception message with the failed address,
to help users figure out what's wrong.

This could potentially be done inside bind() or listen() themselves, but
at that point we no longer have a socket_address anymore. So instead of
providing a different signature to them or even playing with sockaddr ->
socket_address conversions, I just kept the handling in the caller.

Before:

  [shard 0] seastar - Exiting on unhandled exception: std::system_error (error system:98, bind: Address already in use)

After:

  [shard 0] seastar - Exiting on unhandled exception: std::system_error (error system:98, posix_listen failed for address 127.0.0.1:10000: Address already in use)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190304171319.8161-1-glauber@scylladb.com>