core/posix_file_impl: enforce IOV_MAX on vectored IO
The size of the iovec array used by readv() and writev() is limited to
IOV_MAX. This means that most of the users of the vectored IO would have
to ensure that limit is not exceeded in order to avoid EINVAL.

This patch makes posix_file_impl::{read,write}_dma() enforce the limit by
trimming too long vectors of iovecs in a way that preserves the alignment
required by disk operations. The postconditions already allow partial
operations, so the user code already should have been prepared to handle
such situations.