build: Forward variables to CMake correctly
Because the CMake-equivalent names were not specified correctly in
`configure.py`, the wrong CMake variables were being set, and were
subsequently ignored by the configuration process.

We also change the names of the storage variables in `configure.py` to
correspond directly to the CMake names.

We leave the option names themselves unchanged in the interest of
maintaining compatibility.

Fixes #616

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <bbbf3dd4d7ce4708b232d02262a4fb8841692268.1552334815.git.jhaberku@scylladb.com>