core: remove unneeded 'exceptional future ignored' report
In list_directory() function, seastar::stream API is used
to apply a function to each directory entry.
In case the visitor throws, the stream will correctly propagate
the error that happens in stream::produce() and it will be there
when stream::done() is called.

But, the same error still exists in the continuation chain that starts
with stream::started() and it's used to break the do_until loop,
but then it's ultimately ignored. To remedy that, an empty 'handle_exception'
is appended to the chain - which will consume the exceptional future
and get rid of the unneeded log warning.
Message-Id: <3d8bb30c91576fb92d83e9ec4232b8f27fc4cb62.1552407890.git.sarna@scylladb.com>